requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.4
djangorestframework==3.14.0
PyMySQL==1.0.2
pytz==2022.6
sqlparse==0.4.3
